backend/requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
scikit-learn>=1.3.0
joblib>=1.3.0
nltk>=3.8.0
whitenoise>=6.5.0
gunicorn>=21.2.0
